requirements.txt
ccxt==4.3.93
python-dotenv==1.0.0
pandas==1.5.3
numpy==1.24.3
websocket-client==1.3.3